raikiri-style: font-weight computed value を f32 に格上げ (bd raikiri-spike-e52s)
FontWeightValue::Absolute の payload と ComputedValues.font_weight (+
SpecifiedValues.font_weight staging) を u16 から f32 に格上げし、CSS Fonts 4
§2.2.2 "Fractional weights are valid" どおり fractional weight を丸めずに
computed value まで運ぶ。旧実装は parse 段で round-half-away-from-zero によ
り整数化しており、その丸めが §2.2.1 relative-weight table の行選択を変える
2 次被害 (349.5/549.5/749.5 の 3 failure scenario) を伴う既知 divergence だ
った。f32 格上げにより丸め自体が不要になり解消する。

波及: crates/raikiri-dom/src/layout.rs の cv.font_weight as f32 キャストを
削除 (f32 → f32 で clippy::unnecessary_cast に抵触するため)。

FontWeightValue は payload が f32 になったため Eq を derive できなくなり
PartialEq のみに変更 (f32: !Eq)。

Test: 349.5/549.5/749.5 の 3 failure scenario を
bolder_lighter_resolve_against_unrounded_fractional_parent_weight で pin、
WPT font-weight-computed.html の 150.25 computed pin を parse 側 / cascade
側それぞれで追加、既存 font_weight 系 test 全てを f32 literal に更新。

4 gate check 実施:
- cargo fmt --all --check: clean
- cargo clippy --workspace --all-targets -- -D warnings: clean
- cargo test --workspace: 全 pass (raikiri-style 571 tests 含む)
- RUSTDOCFLAGS="-D warnings" cargo doc --no-deps --workspace: raikiri crate
  (未touch の page_scene.rs/entries.rs/page_drawables.rs) で
  private_intra_doc_links エラーが出るが、これは worktree 分岐元 main
  (1239c0c) 時点の pre-existing issue であり、その後 main 上で
  raikiri-spike-6xnf (commit 5c1d3a0) により別途 fix 済み。本 diff が touch
  した raikiri-style / raikiri-dom / raikiri-traits の doc は
  'cargo doc --no-deps -p raikiri-style -p raikiri-dom -p raikiri-traits'
  で個別に clean 確認済み。

Wall handling: この diff は merge しない。wall 判定は report で coordinator
に引き渡す (wall/umbrella 相当の判定根拠あり、詳細は task report 参照)。

+++ crates/raikiri-dom/src/layout.rs
@@ -1087,9 +1087,10 @@ pub(crate) fn preshape_text(
         let mut builder = layout_cx.ranged_builder(fonts, &text, 1.0, true);
         builder.push_default(StyleProperty::FontFamily(font_family));
         builder.push_default(StyleProperty::FontSize(font_size_px));
-        builder.push_default(StyleProperty::FontWeight(FontWeight::new(
-            cv.font_weight as f32,
-        )));
+        // `cv.font_weight` は bd raikiri-spike-e52s で `f32` に格上げ済み
+        // (旧 `u16`) — `parley::FontWeight::new` が要求する型そのものなので
+        // cast は不要 (`as f32` を残すと `clippy::unnecessary_cast` に抵触する)。
+        builder.push_default(StyleProperty::FontWeight(FontWeight::new(cv.font_weight)));
         let mut layout: Layout<()> = builder.build(&text);
         layout.break_all_lines(Some(max_advance));
         // API tuning: brief pseudo-code は `align(Some(max_advance), Alignment::Start,
